Fix OLTSWITCH-253/OLTTS-3980 -Fixed General Query to support Service Identifier   up to 32 bits.
svn path=/trunk/; revision=1500

+++ src/fastpath/src/application/switching/mgmd/src/ptin_mgmd_querier.h
@@ -26,7 +26,7 @@
 
 typedef struct ptinMgmdQuerierInfoDataKey_s
 {  
-  uint16      serviceId; //Service Identifier  
+  uint32      serviceId; //Service Identifier  
 }ptinMgmdQuerierInfoDataKey_t ;
 
 typedef struct ptinMgmdL3Querytimer_s
